fix: restructure the project
