+++ packages/react-ui/src/app/builder/step-settings/branch-settings/branch-condition-group.tsx
@@ -14,7 +14,7 @@ import {
   BranchAction,
 } from '@activepieces/shared';
 
-import { TextInputWithMentions } from '../../text-input-with-mentions/text-input-with-mentions';
+import { TextInputWithMentions } from '../../piece-properties/text-input-with-mentions/text-input-with-mentions';
 
 const textToBranchOperation: Record<BranchOperator, string> = {
   [BranchOperator.TEXT_CONTAINS]: '(Text) Contains',
@@ -77,7 +77,7 @@ const BranchSingleCondition = ({
               <TextInputWithMentions
                 placeholder="First value"
                 onChange={field.onChange}
-                originalValue={field.value}
+                initialValue={field.value}
               ></TextInputWithMentions>
               <FormMessage />
             </FormItem>
@@ -107,7 +107,7 @@ const BranchSingleCondition = ({
                 <TextInputWithMentions
                   placeholder="Second value"
                   {...field}
-                  originalValue={field.value}
+                  initialValue={field.value}
                   onChange={field.onChange}
                 ></TextInputWithMentions>
                 <FormMessage />
