+++ packages/react-ui/src/app/builder/step-settings/step-settings-container.tsx
@@ -15,8 +15,7 @@ import { ScrollArea } from '@/components/ui/scroll-area';
 import { UNSAVED_CHANGES_TOAST, useToast } from '@/components/ui/use-toast';
 import { PieceCardInfo } from '@/features/pieces/components/piece-card-info';
 import { piecesHooks } from '@/features/pieces/lib/pieces-hook';
-import { ActionErrorHandlingForm } from '@/features/properties-form/components/action-error-handling';
-import { formUtils } from '@/features/properties-form/lib/form-utils';
+import { formatUtils } from '@/lib/utils';
 import {
   Action,
   ActionType,
@@ -27,6 +26,8 @@ import {
   flowHelper,
 } from '@activepieces/shared';
 
+import { ActionErrorHandlingForm } from '../piece-properties/action-error-handling';
+import { formUtils } from '../piece-properties/form-utils';
 import { SidebarHeader } from '../sidebar-header';
 import { TestActionComponent } from '../test-step/test-action';
 
@@ -59,8 +60,8 @@ const StepSettingsContainer = React.memo(
 
     const [actionOrTriggerName, setActionOrTriggerName] = useState<string>(
       selectedStep?.settings?.actionName ??
-      selectedStep?.settings?.triggerName ??
-      '',
+        selectedStep?.settings?.triggerName ??
+        '',
     );
 
     const { stepMetadata } = piecesHooks.useStepMetadata({
@@ -103,7 +104,7 @@ const StepSettingsContainer = React.memo(
         actionOrTriggerName,
       },
       resolver: (values, context, options) => {
-        const formSchema = formUtils.buildPieceSchema(
+        const formSchema = formatUtils.buildPieceSchema(
           context.selectedStep.type,
           context.actionOrTriggerName,
           context.pieceModel,
@@ -163,8 +164,8 @@ const StepSettingsContainer = React.memo(
       const currentStep = JSON.parse(JSON.stringify(form.getValues()));
       setActionOrTriggerName(
         currentStep.settings.actionName ??
-        currentStep.settings.triggerName ??
-        '',
+          currentStep.settings.triggerName ??
+          '',
       );
       const newValue = formUtils.buildPieceDefaultValue(
         currentStep,
@@ -228,17 +229,17 @@ const StepSettingsContainer = React.memo(
                   {[ActionType.CODE, ActionType.PIECE].includes(
                     modifiedStep.type as ActionType,
                   ) && (
-                      <ActionErrorHandlingForm
-                        hideContinueOnFailure={
-                          modifiedStep.settings.errorHandlingOptions
-                            ?.continueOnFailure?.hide
-                        }
-                        hideRetryOnFailure={
-                          modifiedStep.settings.errorHandlingOptions
-                            ?.retryOnFailure?.hide
-                        }
-                      ></ActionErrorHandlingForm>
-                    )}
+                    <ActionErrorHandlingForm
+                      hideContinueOnFailure={
+                        modifiedStep.settings.errorHandlingOptions
+                          ?.continueOnFailure?.hide
+                      }
+                      hideRetryOnFailure={
+                        modifiedStep.settings.errorHandlingOptions
+                          ?.retryOnFailure?.hide
+                      }
+                    ></ActionErrorHandlingForm>
+                  )}
                 </div>
               </ScrollArea>
             </ResizablePanel>
