+++ packages/react-ui/src/app/connections/basic-secret-connection-settings.tsx
