+++ packages/react-ui/src/app/connections/connection-table.tsx
@@ -18,7 +18,7 @@ import { authenticationSession } from '@/lib/authentication-session';
 import { formatUtils } from '@/lib/utils';
 import { AppConnection, AppConnectionStatus } from '@activepieces/shared';
 
-import { appConnectionUtils } from '../lib/app-connections-utils';
+import { appConnectionUtils } from '../../features/connections/lib/app-connections-utils';
 
 import { NewConnectionTypeDialog } from './new-connection-type-dialog';
