+++ packages/react-ui/src/app/connections/custom-auth-connection-settings.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { AutoPropertiesFormComponent } from '@/features/properties-form/components/auto-properties-form';
+import { AutoPropertiesFormComponent } from '@/app/builder/piece-properties/auto-properties-form';
 import { CustomAuthProperty } from '@activepieces/pieces-framework';
 
 type CustomAuthConnectionSettingsProps = {
