+++ packages/react-ui/src/app/connections/secret-text-connection-settings.tsx
