+++ packages/react-ui/tsconfig.json
@@ -5,6 +5,7 @@
     "esModuleInterop": false,
     "allowSyntheticDefaultImports": true,
     "strict": true,
+    "noEmit": true,
     "types": ["vite/client"]
   },
   "files": [],
